--- .gitignore
@@ -1,4 +1,0 @@
-*.py[cod]
-secrets.py
-*.log
-.idea
