+++ gamebot/bots/irc_simple.py
