+++ gamebot/coup/exceptions.py
@@ -1,0 +1,25 @@
+__author__ = "jswaro"
+
+
+class CoupException(Exception):
+    pass
+
+
+class MalformedCLICommand(CoupException):
+    pass
+
+
+class InvalidCLICommand(CoupException):
+    pass
+
+
+class GameNotFoundException(CoupException):
+    pass
+
+
+class GamePermissionError(CoupException):
+    pass
+
+
+class GameInvalidOperation(CoupException):
+    pass
