+++ gamebot/coup/player.py
@@ -1,0 +1,45 @@
+__author__ = 'jswaro'
+
+from gamebot.coup.exceptions import GameInvalidOperation
+
+
+class Player(object):
+    def __init__(self, name):
+        self.name = name
+        self.coins = 2
+
+        self.available_influence = list()
+        self.revealed_influence = list()
+
+    def give_card(self, card):
+        self.available_influence.append(card)
+
+    def describe(self):
+        return "{0} has {1} coins".format(self.name, self.coins)
+
+    def current_cards(self):
+        return [x.short_description() for x in self.available_influence] + \
+               [x.short_description() for x in self.revealed_influence]
+
+    def face_up_cards(self):
+        return list(self.revealed_influence)
+
+    def influence_remaining(self):
+        return len(self.available_influence)
+
+    def kill(self):
+        for x in self.available_influence:
+            t = self.available_influence.pop()
+            self.revealed_influence.append(t)
+
+    def cash(self):
+        return self.coins
+
+    def dead(self):
+        return len(self.available_influence) == 0
+
+    def modify_cash(self, amount):
+        if amount + self.coins < 0:
+            raise GameInvalidOperation("You do not have enough coins to perform that action")
+
+        self.coins += amount
