+++ setup.py
@@ -1,0 +1,11 @@
+__author__ = 'jswaro'
+
+from distutils.core import setup
+
+setup(name='Gamebot',
+      version='0.0.1',
+      description='Bot for hosting Coup games',
+      author='James Swaro, Dan Colestock',
+      author_email='james.swaro@gmail.com',
+      url='',
+      packages=['gamebot'])
